Fix some integer values being too small (according to Xcode) on 64-bit hardware.
git-svn-id: https://pcsxr.svn.codeplex.com/svn/pcsxr@85930 e17a0e51-4ae3-4d35-97c3-1a29b211df97

+++ plugins/dfnet/dfnet.h
@@ -75,7 +75,7 @@ Config conf;
 void LoadConf();
 void SaveConf();
 
-long sock;
+int sock;
 char *PadSendData;
 char *PadRecvData;
 char PadSendSize;
@@ -100,7 +100,7 @@ int sockPing();
 int ShowPauseDlg();
 void SysMessage(const char *fmt, ...);
 
-int SEND(const void *pData, int Size, int Mode);
-int RECV(void *pData, int Size, int Mode);
+size_t SEND(const void *pData, int Size, int Mode);
+size_t RECV(void *pData, int Size, int Mode);
 
 #endif
